// js/stat-evaluation-player/src/generated/TouchLastTouchEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RemoteIdTs } from "./RemoteIdTs.ts";

export type TouchLastTouchEvent = { time: number, frame: number, sample_time: number, sample_frame: number, is_team_0: boolean, player: RemoteIdTs | null, player_position?: [number, number, number] | null, };