// js/stat-evaluation-player/src/generated/TouchStatsEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RemoteIdTs } from "./RemoteIdTs.ts";

export type TouchStatsEvent = { time: number, frame: number, sample_time: number, sample_frame: number, player: RemoteIdTs, is_team_0: boolean, kind: string, height_band: string, surface: string, dodge_state: string, ball_speed_change: number, };